Hotfix/cpu pressure (#15)
* fix(pty): cpu pressure

+++ tty_aiohttp/app/handlers/ws/pty.py
@@ -11,13 +11,22 @@ from logging import getLogger
 
 from aiomisc.thread_pool import threaded
 from wsrpc_aiohttp import Route, WSRPCBase, decorators
+from wsrpc_aiohttp.websocket.abc import ProxyMethod
 
 log = getLogger(__name__)
 
 
-@threaded
-def pty_open() -> tuple[int, int]:
-    return pty.openpty()
+@dataclass
+class PTYConfig:
+    master_fd: int
+    slave_fd: int
+    shell: str
+
+    @classmethod
+    @threaded
+    def open_pty(cls, shell: str) -> "PTYConfig":
+        master_fd, slave_fd = pty.openpty()
+        return cls(master_fd, slave_fd, shell)
 
 
 @dataclass
@@ -26,29 +35,37 @@ class Terminal:
     fd: int
     read_queue: asyncio.Queue[str]
     write_queue: asyncio.Queue[str]
-
-    read_cb: t.Callable[[str], t.Awaitable[None]]
+    proxy: ProxyMethod
 
     _read_task: asyncio.Task[None] = field(init=False)
+    _write_task: asyncio.Task[None] = field(init=False)
+    _monitor_task: asyncio.Task[None] = field(init=False)
 
     def __post_init__(self) -> None:
-        self._read_task = asyncio.create_task(self.read())
+        """Initialize terminal tasks"""
+        self._write_task = asyncio.create_task(self._write())
+        self._read_task = asyncio.create_task(self._read())
+        self._monitor_task = asyncio.create_task(self._monitor())
         loop = asyncio.get_event_loop()
         loop.add_reader(self.fd, self.on_read)
-        loop.add_writer(self.fd, self.on_write)
 
-    async def read(self) -> None:
+    async def _read(self) -> None:
         while True:
             chunk = await self.read_queue.get()
-            await self.read_cb(chunk)
+            await self.proxy.output(data=chunk)
+
 
     async def write(self, chunk: str) -> None:
         await self.write_queue.put(chunk)
 
-    def on_write(self) -> None:
-        while not self.write_queue.empty():
-            user_input: str = self.write_queue.get_nowait()
-            os.write(self.fd, user_input.encode("utf-8"))
+    async def _write(self) -> None:
+        while True:
+            chunk = await self.write_queue.get()
+            await self._do_write(chunk.encode("utf-8"))
+
+    @threaded
+    def _do_write(self, chunk: bytes) -> None:
+        os.write(self.fd, chunk)
 
     def on_read(self) -> None:
         max_read_bytes = 1024 * 20
@@ -66,15 +83,32 @@ class Terminal:
         winsize = struct.pack("HHHH", height, width, xpix, ypix)
         fcntl.ioctl(self.fd, termios.TIOCSWINSZ, winsize)
 
-    async def close(self) -> None:
-        self._read_task.cancel()
-        self.process.kill()
-        status_code = await self.process.wait()
-        log.debug("Terminal closed with status code %s", status_code)
+    async def _monitor(self) -> None:
+        return_code = await self.process.wait()
+        log.info("Process was finished with return code %s", return_code)
+        await self._close()
 
+    async def _close(self) -> None:
         loop = asyncio.get_event_loop()
         loop.remove_reader(self.fd)
-        loop.remove_writer(self.fd)
+        self._read_task.cancel()
+        self._write_task.cancel()
+        message = f"""Shell was closed
+with return code {self.process.returncode}"""
+        await self.proxy.notify(
+            type="error",
+            title="Terminal is closed",
+            message=message,
+        )
+
+    async def close(self) -> None:
+        if (return_code := self.process.returncode) is not None:
+            log.info(
+                "Process was already closed with return code %s",
+                return_code,
+            )
+            return
+        self.process.kill()
 
 
 class PtyHandler(Route):
@@ -90,40 +124,36 @@ class PtyHandler(Route):
     async def terminal(self) -> Terminal:
         if self._terminal is not None:
             return self._terminal
-        master_fd, slave_fd = await pty_open()
+        pty_config = await PTYConfig.open_pty(self.shell)
         process = await asyncio.create_subprocess_exec(
             self.shell,
             preexec_fn=os.setsid,
-            stdin=slave_fd,
-            stdout=slave_fd,
-            stderr=slave_fd,
+            stdin=pty_config.slave_fd,
+            stdout=pty_config.slave_fd,
+            stderr=pty_config.slave_fd,
             env=os.environ.copy(),
         )
+        socket: WSRPCBase = self.socket  # type: ignore
         terminal = Terminal(
-            process, master_fd, asyncio.Queue(), asyncio.Queue(), self.on_read
+            process, pty_config.master_fd, asyncio.Queue(), asyncio.Queue(),
+            socket.proxy.pty,
         )
 
         self._terminal = terminal
         return self._terminal
 
-    async def on_read(self, data: str | bytes) -> None:
-        socket: WSRPCBase = self.socket  # type: ignore
-        await socket.proxy.pty.output(data=data)
-
     @decorators.proxy
-    async def ready(self) -> None:
-        await self.terminal
+    async def ready(self, width: int, height: int) -> None:
+        await self.resize(width=width, height=height)
 
     @decorators.proxy
     async def input(self, data: str) -> None:
         terminal = await self.terminal
-
         await terminal.write(data)
 
     @decorators.proxy
     async def resize(self, width: int, height: int) -> None:
         terminal = await self.terminal
-
         await terminal.resize(width=width, height=height)
 
     async def _onclose(self) -> None:  # type: ignore
